+++ src/open_bos_stream/web_access/__init__.py
@@ -1,0 +1,1 @@
+"""Optionaler Webzugriff über den HTTP-Standardport."""